refactor(voice/capture): extract session-manager contention helpers [T1.4 step 3]
Master mission Phase 1 / T1.4 — third stage of the
``_capture_task.py`` split. Builds on step 1 (`32efb40`,
restart-verdict types) and step 2 (`c960f6c`, exception
hierarchy) by extracting the Linux session-manager contention
pattern classifier. Same low-risk profile: pure functions, no
class state coupling.

What lands:

  * New module ``capture/_contention.py`` containing:
      - ``_SESSION_MANAGER_CONTENTION_ERROR_CODES`` —
        frozenset of ErrorCode values interpreted as "another
        client holds the device".
      - ``_is_session_manager_contention_pattern`` — Linux-
        specific classifier returning True iff every open
        attempt fell in the contention class. Pure function.
      - ``_suggest_session_manager_alternatives`` — UI-facing
        action token list. Pure function.
    Total ~75 LOC. ``OpenAttempt`` typed-only import preserved
    via ``TYPE_CHECKING``.
  * ``_capture_task.py`` drops the three definitions, adds
    ``import X as X`` re-exports for each name. ``test_capture_device_contended_error.py``
    imports two of the helpers directly from the legacy path; the
    re-export keeps that test working unchanged.
  * ``capture/__init__.py`` re-exports the three names alongside
    the step-1 + step-2 surfaces.

T1.4 cumulative impact (steps 1-3):

  * ``_capture_task.py``: 2785 → 2313 LOC (~470 LOC extracted,
    ~17 % of the split done; cumulative drop ~17 %).
  * Three coherent sub-modules under ``voice/capture/``:
    ``_restart.py`` (verdict types), ``_exceptions.py`` (error
    hierarchy), ``_contention.py`` (session-manager classifier).
    All pure-types / pure-functions content with zero class
    state coupling.
  * Remaining: the ``AudioCaptureTask`` class itself (~1972 LOC)
    + module-level constants + ``_rms_db_int16`` / ``_PEAK_DB_RE``
    / ``_extract_peak_db`` / ``_resolve_input_entry`` low-level
    helpers. The ``AudioCaptureTask`` class via the
    ``_RestartMixin`` pattern remains the genuinely-risky
    "DEDICATED SESSION" surgery; future steps can land
    independently with the legacy stub intact.

Validation:
  * ruff check + format clean (auto-fix removed an unused
    ``StrEnum`` / ``frozenset`` / ``Sequence`` import that the
    extraction left dangling)
  * mypy strict: 0 issues across 422 source files (+1 source
    file for ``capture/_contention.py``)
  * bandit clean
  * pytest full suite: 11894 passed, 25 skipped, 57 deselected
    (zero delta — pure structural extraction, no test changes
    needed)

Mission: docs-internal/missions/MISSION-voice-final-skype-grade-2026.md §Phase 1.T1.4 (step 3)
