+++ src/sovyx/voice/_capture_task.py
@@ -58,6 +58,22 @@ from sovyx.voice._chaos import ChaosInjector, ChaosSite
 from sovyx.voice._frame_normalizer import FrameNormalizer
 from sovyx.voice._stream_opener import _import_sounddevice
 
+# T1.4 step 3 — Linux session-manager contention helpers extracted to
+# ``voice/capture/_contention``. Re-exported via the explicit
+# ``import X as X`` pattern so existing imports — particularly
+# ``test_capture_device_contended_error.py`` which imports the two
+# public helpers directly — keep working without an import-path
+# migration.
+from sovyx.voice.capture._contention import (
+    _SESSION_MANAGER_CONTENTION_ERROR_CODES as _SESSION_MANAGER_CONTENTION_ERROR_CODES,
+)
+from sovyx.voice.capture._contention import (
+    _is_session_manager_contention_pattern as _is_session_manager_contention_pattern,
+)
+from sovyx.voice.capture._contention import (
+    _suggest_session_manager_alternatives as _suggest_session_manager_alternatives,
+)
+
 # T1.4 step 2 — exception class hierarchy extracted to
 # ``voice/capture/_exceptions``. Re-exported via the explicit
 # ``import X as X`` pattern so existing imports like
@@ -124,13 +140,12 @@ from sovyx.voice.capture._restart import (
 )
 
 if TYPE_CHECKING:
-    from collections.abc import Callable, Sequence
+    from collections.abc import Callable
 
     import numpy as np
     import numpy.typing as npt
 
     from sovyx.engine.config import VoiceTuningConfig
-    from sovyx.voice._stream_opener import OpenAttempt
     from sovyx.voice.device_enum import DeviceEntry
     from sovyx.voice.pipeline._orchestrator import VoicePipeline
 
@@ -231,80 +246,6 @@ fire one WARN per consumer-loop iteration, drowning the dashboard.
 30 s matches the typical operator response cadence — long enough
 that a recovering condition self-suppresses, short enough that an
 unattended outage produces a regular drumbeat in the log feed."""
-
-
-# ── T7 session-manager contention helpers ────────────────────────────
-
-_SESSION_MANAGER_CONTENTION_ERROR_CODES: frozenset[str] = frozenset(
-    {
-        "device_busy",
-        "device_disappeared",
-        "device_not_found",
-    }
-)
-"""ErrorCode values interpreted as "another client holds the device".
-
-PortAudio on Linux returns ``-9985 Device unavailable`` for the common
-"PipeWire grabbed hw:X,Y" pathology. The opener classifies that as
-``ErrorCode.DEVICE_BUSY``. ``DEVICE_DISAPPEARED`` covers the related
-``-9988 Device disappeared`` and ``DEVICE_NOT_FOUND`` is included
-because some kernel-invalidated states surface as ``-9996 Invalid
-device`` when a session manager yanks the exclusive lock mid-open.
-"""
-
-
-def _is_session_manager_contention_pattern(
-    *,
-    platform: str,
-    open_attempts: Sequence[OpenAttempt],
-) -> bool:
-    """Return ``True`` iff the attempt list matches "session manager holds hw".
-
-    The rule is intentionally narrow — false positives would only
-    swap a generic ``RuntimeError`` message for a slightly more useful
-    one (no regression risk), but we still constrain the heuristic to
-    (a) Linux only, (b) at least one attempt made, (c) every attempt
-    falls in the contention-class :data:`_SESSION_MANAGER_CONTENTION_ERROR_CODES`.
-
-    The ``attempts_tried_hw_and_virtual`` half of the ADR rule is
-    handled upstream by the candidate-set: when this function fires,
-    the opener already iterated the opener-side pyramid on the *current*
-    candidate, and the cascade-level loop in
-    :func:`~sovyx.voice.health.cascade.run_cascade_for_candidates` has
-    exhausted every candidate (hardware + virtual). Re-checking here
-    would require access to the cascade history, which the capture
-    task legitimately does not have. Keeping the check at open-level
-    is sound because the cascade only reaches ``start()`` on a device
-    it already considered "best bet remaining" — a device-busy cluster
-    at this stage implies every earlier candidate also failed.
-    """
-    if platform != "linux":
-        return False
-    if not open_attempts:
-        return False
-    return all(
-        attempt.error_code is not None
-        and attempt.error_code.value in _SESSION_MANAGER_CONTENTION_ERROR_CODES
-        for attempt in open_attempts
-    )
-
-
-def _suggest_session_manager_alternatives() -> list[str]:
-    """Return the UI-facing action tokens for a session-manager grab.
-
-    Order: preferred alternative first. The dashboard maps each token
-    to an i18n key + an action (chip click dispatches the corresponding
-    fallback request). Currently static — future revisions may query
-    enumeration to elide tokens for devices that don't exist on the
-    host, but doing so here would introduce a sync ``sounddevice`` call
-    on the error path.
-    """
-    return [
-        "select_device:pipewire",
-        "select_device:default",
-        "select_device:pulse",
-        "stop_process:pipewire",
-    ]
 
 
 def _rms_db_int16(frame: Any) -> float:  # noqa: ANN401 — numpy int16 array; Any keeps numpy lazy-imported
